fix: address follow-up proxy auth review
