+++ tools/cfl/integration-tests.md
@@ -758,6 +758,27 @@ All cfl commands work with all configured auth methods (no scope restrictions fo
 
 ---
 
+### Pass 3: Proxy Auth
+
+#### Setup (Proxy Auth)
+- [ ] Configure a trusted proxy that authenticates upstream and exposes a loopback `http://` URL or an `https://` URL
+- [ ] `cfl init --auth-method proxy --url http://127.0.0.1:8080/atlassian --no-verify` (adjust URL for your proxy)
+- [ ] `cfl config show` — auth_method = proxy; email, api_token, and cloud_id are empty
+- [ ] `cfl config test` — Connection verified through the proxy
+- [ ] `cfl space list` works
+
+#### Full Checklist (Proxy Auth)
+- [ ] Run Page CRUD, ADF Body Fallback, Wiki Links, Attachment CRUD, Search, Space CRUD, and Edge Cases using the same steps as Pass 1
+- [ ] No sections are skipped; proxy auth has no cfl scope restrictions
+
+#### Cleanup (Proxy Auth)
+- [ ] Delete all [Test] prefixed pages
+- [ ] `cfl space delete INTTEST --force`
+- [ ] `cfl space delete INTTEST2 --force`
+- [ ] Verify no test data remains
+
+---
+
 ## Adding New Tests
 
 When adding new features or fixing bugs:
@@ -765,4 +786,4 @@ When adding new features or fixing bugs:
 1. Add test cases to the appropriate section above
 2. Include both happy path and error cases
 3. Document any known limitations or edge cases
-4. Update the "Test Execution Checklist" if needed
+4. Update Pass 1, Pass 2, and Pass 3 in the Test Execution Checklist
