+++ tools/jtk/internal/cmd/initcmd/initcmd_test.go
@@ -68,86 +68,6 @@ func TestRunInit_InvalidAuthMethod(t *testing.T) {
 // The non-interactive paths (all flags provided) still use huh forms internally,
 // so we test config loading/saving separately
 
-// TestRequireNonInteractiveFields_NamesFirstMissing pins the §3.4
-// fail-loud message shape so the family pattern (jtk + cfl + future
-// nrq-aligned ports) stays consistent. The wizard wrapper isn't tested
-// directly because it depends on huh form state we can't easily fake
-// — the helper IS the contract for what gets named.
-func TestRequireNonInteractiveFields_NamesFirstMissing(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name  string
-		cfg   *config.Config
-		wants []string
-	}{
-		{
-			name:  "basic auth — missing URL",
-			cfg:   &config.Config{},
-			wants: []string{"--url"},
-		},
-		{
-			name:  "basic auth — missing email",
-			cfg:   &config.Config{URL: "https://acme.atlassian.net"},
-			wants: []string{"--email"},
-		},
-		{
-			name:  "bearer — missing cloud-id",
-			cfg:   &config.Config{URL: "https://acme.atlassian.net", AuthMethod: auth.AuthMethodBearer},
-			wants: []string{"--cloud-id"},
-		},
-		{
-			name:  "basic auth — missing token recommends --token-stdin + --token-from-env + set-credential",
-			cfg:   &config.Config{URL: "https://acme.atlassian.net", Email: "u@x.io"},
-			wants: []string{"--token-stdin", "--token-from-env", "set-credential"},
-		},
-		{
-			name:  "bearer — missing token recommends --token-stdin + --token-from-env + set-credential",
-			cfg:   &config.Config{URL: "https://acme.atlassian.net", CloudID: "cid", AuthMethod: auth.AuthMethodBearer},
-			wants: []string{"--token-stdin", "--token-from-env", "set-credential"},
-		},
-	}
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			err := requireNonInteractiveFields(tc.cfg)
-			testutil.RequireError(t, err)
-			if !strings.Contains(err.Error(), "--non-interactive: missing") {
-				t.Fatalf("error must mention --non-interactive prefix: %v", err)
-			}
-			for _, want := range tc.wants {
-				if !strings.Contains(err.Error(), want) {
-					t.Fatalf("error must name %s, got %v", want, err)
-				}
-			}
-		})
-	}
-}
-
-// TestRequireNonInteractiveFields_AllSupplied_NoError — happy path; no
-// error returned when every required field is present.
-func TestRequireNonInteractiveFields_AllSupplied_NoError(t *testing.T) {
-	t.Parallel()
-	cfg := &config.Config{
-		URL: "https://acme.atlassian.net", Email: "u@x.io",
-		APIToken: "tok-1234567890",
-	}
-	if err := requireNonInteractiveFields(cfg); err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
-
-func TestRequireNonInteractiveFields_ProxyNeedsOnlyURL(t *testing.T) {
-	t.Parallel()
-	cfg := &config.Config{
-		URL:        "http://127.0.0.1:8080/atlassian",
-		AuthMethod: auth.AuthMethodProxy,
-	}
-	if err := requireNonInteractiveFields(cfg); err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
-
 // TestRunInit_NonInteractive_MissingURL_Fails — drives runInit through
 // the public surface with --non-interactive but no flags supplied. The
 // fail-loud message must surface BEFORE any keyring/migration work runs.
